Switch from spacemacs to custom emacs config!!
I found that spacemacs has become a bit of a kitchen sink, and I'm
tired of not grokking how all the different bits are fitting
together. Particularly because enabling things like company-mode and
flycheck would invariably lead to emacs going out to lunch all the
time in the haskell and rust modes. It was helpful for finding out
about lots of nice emacs stuff, though.